update Order and OrderDetail

+++ src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -20,7 +20,7 @@ public class OrderDetailDTO {
     private Long productId;
 
     @Min(value = 0, message = "Price must be >= 0")
-    private Long price;
+    private Float price;
 
     @JsonProperty("number_of_products")
     @Min(value = 1, message = "Number of products must be >= 1")
@@ -28,7 +28,7 @@ public class OrderDetailDTO {
 
     @JsonProperty("total_money")
     @Min(value = 0, message = "Total money must be >= 0")
-    private Long totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
